main/alpine-baselayout: ensure default PS1 is set to something reasonable
We continue to export PS1 as this would otherwise be a breaking change.  The
question of whether or not this behavior should be changed can be resolved
through the standard system change proposal process.

We do however use a simplified version of the PS1 prompt generation based on
the proposal made by mirabilos in #12398, without using $ENV.  If somebody
wishes to pursue the full proposal by mirabilos, this can also be done through
the standard system change proposal process.

Closes #12398.